+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - user: add simple empty states
 - format joined date
 - display preferred user license
+- index(search): autoFocus input
 
 ## [0.1.1] - 2021-04-07
 ## Features